move stx submit hook
